feat: 商户信息

+++ vscing-model/src/main/java/com/vscing/model/entity/MerchantBill.java
@@ -36,12 +36,12 @@ public class MerchantBill extends BaseEntity {
   private BigDecimal changeAfterBalance;
 
   @Schema(description = "商品类型 1 电影票")
-  private Short productType;
+  private Integer productType;
 
   @Schema(description = "变动类型 1 订单支付扣款 2 订单退款 3 账户充值 4 账户退款")
-  private Short changeType;
+  private Integer changeType;
 
   @Schema(description = "状态 1 进行中 2 成功 3 失败")
-  private Short status;
+  private Integer status;
 
 }
